==> 7.3 MBA Approach 3.ipynb <==
# Market Basket analysis using ranking

## Import Libraries

# Import modules that are used in this notebook. 
# Some additions would be there later on depending on when a need arises

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
%matplotlib inline
import seaborn as sns

pd.pandas.set_option('display.max_columns',None)

## Read data

#read the original data file and understand the data
dataset_orig=pd.read_csv('Merged_clean.csv')
print(dataset_orig.shape)

## Get subset of only multi policy holders

#from original dataset get all those records where customers have more then one policy
df_multi = dataset_orig[dataset_orig['Freq'] > 1].sort_values(['policy_owner_number','RCD'])

## within each group give a rank order

df_multi['rank'] = df_multi.groupby(['policy_owner_number']).cumcount()+1
df_multi.info()

#display the data
print('\033[1m Product_Description \033[0m : \n',df_multi[df_multi['rank']==1]['Product_Description'].value_counts(normalize=True).head(5))
print('#########################')

print('\033[1m Product_brief_category \033[0m : \n',df_multi[df_multi['rank']==1]['Product_brief_category'].value_counts(normalize=True).head(5))
print('#########################')

print('\033[1m Product_Club_Manual \033[0m : \n',df_multi[df_multi['rank']==1]['Product_Club_Manual'].value_counts(normalize=True).head(5))
print('#########################')

print('\033[1m CUST_prod_cat \033[0m : \n',df_multi[df_multi['rank']==1]['CUST_prod_cat'].value_counts(normalize=True).head(5))
print('#########################')

print('\033[1m Par_NonPar \033[0m : \n',df_multi[df_multi['rank']==1]['Par_NonPar'].value_counts(normalize=True).head(5))
print('#########################')

## Find the customers who have bought the top first policy from the list

# so if first policy is guranteed income, get the counts of the second bought policy
cust_ids_Product_Description = df_multi[(df_multi['rank']==1)& 
                    (df_multi['Product_Description'] == 'GURANTEED INCOME')  ]['policy_owner_number']

df_multi_2_Product_Description = df_multi[(df_multi['rank']==2) &
                     (df_multi.policy_owner_number.isin(cust_ids_Product_Description))]


cust_ids_Product_brief_category = df_multi[(df_multi['rank']==1)& 
                    (df_multi['Product_brief_category'] == 'TRADITIONAL')  ]['policy_owner_number']

df_multi_2_Product_brief_category = df_multi[(df_multi['rank']==2) &
                     (df_multi.policy_owner_number.isin(cust_ids_Product_brief_category))]

cust_ids_Product_Club_Manual = df_multi[(df_multi['rank']==1)& 
                    (df_multi['Product_Club_Manual'] == 'GUARANTEED INCOME')  ]['policy_owner_number']
df_multi_2_Product_Club_Manual = df_multi[(df_multi['rank']==2) &
                     (df_multi.policy_owner_number.isin(cust_ids_Product_Club_Manual))]

cust_ids_CUST_prod_cat = df_multi[(df_multi['rank']==1)& 
                    (df_multi['CUST_prod_cat'] == 'TRADITIONAL')  ]['policy_owner_number']
df_multi_2_CUST_prod_cat = df_multi[(df_multi['rank']==2) &
                     (df_multi.policy_owner_number.isin(cust_ids_CUST_prod_cat))]

cust_ids_Par_NonPar = df_multi[(df_multi['rank']==1)& 
                    (df_multi['Par_NonPar'] == 'PAR')  ]['policy_owner_number']
df_multi_2_Par_NonPar = df_multi[(df_multi['rank']==2) &
                     (df_multi.policy_owner_number.isin(cust_ids_Par_NonPar))]

## Check the highest

print(df_multi_2_Product_Description['Product_Description'].value_counts(normalize=True).head(1))
print(df_multi_2_Product_brief_category['Product_brief_category'].value_counts(normalize=True).head(1))
print(df_multi_2_Product_Club_Manual['Product_Club_Manual'].value_counts(normalize=True).head(1))
print(df_multi_2_CUST_prod_cat['CUST_prod_cat'].value_counts(normalize=True).head(1))
print(df_multi_2_Par_NonPar['Par_NonPar'].value_counts(normalize=True).head(1))

#    Product_Description       Product_brief_category   Product_Club_Manual        CUST_prod_cat       Par_NonPar
#1   GURANTEED INCOME(23.17%)  TRADITIONAL(87.44%)      GUARANTEED INCOME(23.17%)  TRADITIONAL(69%)    PAR(55.05%)
#2   GURANTEED INCOME(63.14%)  TRADITIONAL(88.63%)      GUARANTEED INCOME(63.14%)  TRADITIONAL(100%)   PAR(68.68%)
